get RegionServer domain before send state
