+++ .gitignore
@@ -43,3 +43,5 @@ data
 /terminal/testrun/
 /terminal/downloadldb/
 /terminal/debug/
+/terminal/v2.1.0
+/terminal/meson
